backend/requirements.txt
Django==3.2
djangorestframework==3.13.1
djoser==2.1.0
Pillow==9.0.1
sorl-thumbnail==12.8.0
